+++ client/src/App.css
@@ -1,6 +1,6 @@
 #root {
   width: 100%;
-  margin: 0 auto;
+  margin: 0;
   padding: 2rem;
   text-align: center;
 }
